+++ src/lib/forms-schema/index.ts
@@ -845,7 +845,6 @@ const formSchema = Joi.object().keys({
           .required(),
       }),
     ),
-  isAIBuilderSupported: Joi.boolean().default(false),
 })
 
 export const formEventTypes: SubmissionEventTypes.FormEventType[] = [